Scroll faster on navbar and adjust when size changes
